requirements.txt
pandas
streamlit
requests
google-generativeai==0.3.2
streamlit_option_menu==0.3.13
PyPDF2==3.0.1